requirements.txt
requests>=2.31.0
pyyaml>=6.0
google-cloud-bigquery>=3.14.0
streamlit>=1.30.0
folium>=0.15.0
streamlit-folium>=0.15.0
pandas>=2.1.0
